Create Rock-Paper-Scissors
This is my first attempt at coding out a game with absolutely no assistance.  I wanted to challenge my knowledge and see if I had been learning from my tutorial classes.  It might not be exactly how I intended it to come out but it works.   Oh the joy of getting a program to function!